Rename default grammar to strict.
And aft to default.

Rather than making new files for the exceptions of each vendor or product,
it's probably easier for the default grammar to be loose and incorporate the
various extensions, changes, idiosyncracies etc.  Then the original grammar
can be renamed to -strict for people who would like that.